@@ -1,5 +1,5 @@
 import React from 'react';
-import {Accounts} from '../components/accounts';
+import {Accounts} from '@/components/accounts';
 
 const accounts = () => {
    return <Accounts />;
